+++ src/App.tsx
@@ -1,19 +1,17 @@
 import { BrowserRouter } from 'react-router-dom';
 import { CssBaseline, ThemeProvider } from '@mui/material';
-import BodyRouting from './components/Layout/BodyRouting';
-import BodyContent from './components/Layout/BodyContent';
+import Layout from './components/Layout/Layout';
+import PartySelection from './pages/partySelection/PartySelection';
 import theme from './theme';
-// const theme = createTheme({});
 
 export function App() {
   return (
     <BrowserRouter>
       <ThemeProvider theme={theme}>
         <CssBaseline />
-        <BodyRouting>
-          <BodyContent/>
-        </BodyRouting>
-        
+        <Layout>
+          <PartySelection/>
+        </Layout>
       </ThemeProvider>
     </BrowserRouter>
   );
